Cache RoPE cos/sin dtype cast instead of recomputing per layer
attention.forward cast the cos/sin slice to the activation dtype on every
layer of every decode step (2 casts x 32 layers = 64 tiny copy kernels per
step). The RoPE tables are constant for the whole run, so cast the full
table once per dtype and slice it. Slicing a cast table is bit-identical to
casting a slice (cast is elementwise), so logits are unchanged (verified
max|delta|=0) while 64 launches per decode step are removed.

+++ ops/attention.py
@@ -128,9 +128,9 @@ class GroupedQueryAttention(nn.Module):
         v = v.view(B, T, self.num_heads_kv, self.head_dim)
 
         # --- 3. Apply RoPE to Q and K ---
-        cos, sin = self.rope_freqs.get(seq_len=T, start_pos=start_pos)
-        cos = cos.to(x.dtype)
-        sin = sin.to(x.dtype)
+        # Fetch cos/sin already in the activation dtype from RopeFrequencies'
+        # cached cast tables (avoids a per-layer, per-step .to(dtype) copy).
+        cos, sin = self.rope_freqs.get(seq_len=T, start_pos=start_pos, dtype=x.dtype)
         q, k = apply_rope(q, k, cos, sin)
 
         # --- 4. Transpose to (B, n_heads, T, head_dim) for SDPA ---
